refactored work environment to factory in seperated files
