+++ tests/test_workenvironment_aml.py
@@ -1,5 +1,5 @@
 import os
-from arcus.azureml.environment.workenvironment import WorkEnvironment
+from arcus.azureml.environment.environment_factory import WorkEnvironmentFactory as fac
 from collections import Counter
 
 def is_interactive():
@@ -18,13 +18,13 @@ datastore_name = 'smart_devops'
 partitioned_datastore_name = 'aiincubators_covid'
 
 def test_dataset():
-    work_env = WorkEnvironment.Create(connected=True)
+    work_env = fac.Create(connected=True)
     _df = work_env.load_tabular_dataset('smart-devops-changesets')
     _df = _df.tail(20)
     assert _df.shape == (20,16) # 16 columns expected
 
 def test_partitions():
-    work_env = WorkEnvironment.Create(connected=True, datastore_path='arcus_partition_test')
+    work_env = fac.Create(connected=True, datastore_path='arcus_partition_test')
     partition_df = work_env.load_tabular_partition(partition_name= 'test-partitioning/stock_AT*', columns=['Close', 'High', 'Isin', 'ItemDate', 'Low', 'Market', 'Open', 'Ticker', 'Volume'])
     assert partition_df is not None
     assert partition_df.shape == (30, 9)
@@ -32,7 +32,7 @@ def test_partitions():
     assert len(Counter(partition_df['Isin'])) == 3
 
 def test_partitions_header():
-    work_env = WorkEnvironment.Create(connected=True, datastore_path='arcus_partition_test')
+    work_env = fac.Create(connected=True, datastore_path='arcus_partition_test')
     partition_df = work_env.load_tabular_partition('test-partitioning/stock_header_AT*', first_row_header=True)
     assert 'Isin' in partition_df.columns
     assert partition_df.shape == (24, 9)
@@ -40,7 +40,7 @@ def test_partitions_header():
 
 
 def test_partitions_noheader_and_columns():
-    work_env = WorkEnvironment.Create(connected=True, datastore_path='arcus_partition_test')
+    work_env = fac.Create(connected=True, datastore_path='arcus_partition_test')
     partition_df = work_env.load_tabular_partition('test-partitioning/stock_header_AT*', first_row_header=False, columns=['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I'])
     assert partition_df is not None
 
@@ -48,7 +48,7 @@ def test_partitions_noheader_and_columns():
     assert partition_df.shape == (27, 9)
 
 def test_partitions_header_and_columns():
-    work_env = WorkEnvironment.Create(connected=True, datastore_path='arcus_partition_test')
+    work_env = fac.Create(connected=True, datastore_path='arcus_partition_test')
     partition_df = work_env.load_tabular_partition('test-partitioning/stock_header_AT*', first_row_header=True, columns=['Close', 'High', 'Isin', 'ItemDate', 'Low', 'Market', 'Open', 'Ticker', 'Volume'])
     assert partition_df is not None
 
@@ -57,6 +57,6 @@ def test_partitions_header_and_columns():
     assert len(Counter(partition_df['Isin'])) == 3
 
 def test_partitions_notexisting():
-    work_env = WorkEnvironment.Create(connected=True, datastore_path='arcus_partition_test')
+    work_env = fac.Create(connected=True, datastore_path='arcus_partition_test')
     partition_df = work_env.load_tabular_partition('test-partitioning/stock_BE*', columns=['Close', 'High', 'Isin', 'ItemDate', 'Low', 'Market', 'Open', 'Ticker', 'Volume'])
     assert partition_df is None
